CanineCompanionBackend/requirements.txt
langchain==0.1.12
openai==1.10.0
scikit-learn==1.3.0
pinecone-client==3.0.2
python-dotenv==1.0.0
scikit-learn==1.3.0